check for /proc/sys/kernel/hotplug before apt-installing hotplug
git-svn-id: svn://svn.debian.org/svn/d-i/trunk@11689 48c42b26-1dd6-0310-b98f-a58d8bce7237

+++ tools/ddetect/hw-detect.sh
@@ -308,7 +308,9 @@ fi
 apt-install discover || true
 
 # Install hotplug as well (for USB, IEEE1394, CardBus, and some SCSI)
-apt-install hotplug || true
+if [ -f /proc/sys/kernel/hotplug ]; then 
+	apt-install hotplug || true
+fi
 
 db_progress SET $MAX_STEPS
 db_progress STOP
